stackstring, tightstring extraction with unicorn engine, move out stringfinder table to a new file, fix some bugs
